Move show_flashcard_selector and FlashCardSelectorDialog to a file flashcard_selector_dialog.py
